To investigate, how to not duplicate the rows when trying to rerun scripts for step01_financials
When try to re-run the scripts to merge data for silver_01 instead of replacing, it duplicates all the  rows
-Investigate bulk_insert_dataframe, if_exists = "append" vs "replace". Append will duplicate, replace will overwrite the data with fresh data.

+++ python scripts/bronze/08_load_balance_sheet_sql.py
@@ -136,7 +136,7 @@ CREATE TABLE {TARGET_SCHEMA_TABLE}
     -- General Info
     --========================================================
     symbol                              VARCHAR(20)     NOT NULL,
-    calendar_year                         SMALLINT,
+    calendar_year                       INT,
 
     --========================================================
     -- Data
