+++ python scripts/bronze/10_load_cash_flow_sql.py
@@ -130,7 +130,7 @@ CREATE TABLE {TARGET_SCHEMA_TABLE}
     -- General Info
     --========================================================
     symbol                  VARCHAR(20)     NOT NULL,
-    calendar_year           SMALLINT,
+    calendar_year           INT,
 
     --========================================================
     -- Data
